move markdown functionality to the pages controller and link the terms
