--- dpc-web/app/controllers/docs_controller.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class DocsController < ApplicationController  
-end
