--- dpc-web/app/helpers/docs_helper.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-module DocsHelper
-end
