+++ dpc-web/app/views/devise/registrations/new.html.erb
@@ -85,7 +85,9 @@
 
         <div class="field ds-u-margin-top--3">
           <%= f.check_box :agree_to_terms, class: 'ds-c-choice' %>
-          <%= f.label :agree_to_terms, 'I agree to the terms service' %>
+          <%= f.label :agree_to_terms do %>
+          I agree to the <%= link_to " terms of service", terms_of_service_path %>
+          <% end %>
         </div>
 
         <div class="actions ds-u-margin-top--3">
