+++ dpc-web/app/views/pages/guide.html.erb
