+++ dpc-web/app/views/pages/guide.md
