+++ dpc-web/config/routes.rb
@@ -12,9 +12,8 @@ Rails.application.routes.draw do
 
   match '/home', to: 'public#home', via: :get
 
-  match '/docs', to: 'docs#reference', via: :get
-  match '/docs/guide', to: 'docs#guide', via: :get
-
+  match '/docs', to: 'pages#reference', via: :get
+  match '/docs/guide', to: 'pages#guide', via: :get
   match '/faq', to: 'pages#faq', via: :get
   match '/support', to: 'pages#support', via: :get
   match '/terms-of-service', to: 'pages#terms_of_service', via: :get
